pyproject.toml
[project]
name = "blog-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.55.2",
    "python-docx>=1.1.2",
]
